+++ src/optigob/static_ag/baseline_static_ag.py
@@ -30,6 +30,9 @@ Methods:
     get_pig_and_poultry_area: Get the area for Pig and Poultry systems.
     get_crop_area: Get the area for Crop systems.
     get_total_static_ag_area: Get the total area for all static agricultural systems.
+    get_sheep_protein: Get the protein value for Sheep systems.
+    get_pig_and_poultry_protein: Get the protein value for Pig and Poultry systems.
+    get_total_static_ag_protein: Get the total protein value for all static agricultural systems.
 """
 
 class BaselineStaticAg:
@@ -309,3 +312,43 @@ class BaselineStaticAg:
         crop_area = self.get_crop_area()
 
         return sheep_area + pig_and_poultry_area + crop_area
+    
+    
+    def get_sheep_protein(self):
+        """
+        Get the protein value for Sheep systems.
+
+        Returns:
+            float: The protein value in kg.
+        """
+        sheep_protein = self.data_manager_class.get_static_livestock_protein_scaler(
+            year=self.baseline_year, system='Sheep', item='meat',abatement=self.abatement_type
+        )
+
+        return sheep_protein["value"].item()
+    
+
+    def get_pig_and_poultry_protein(self):
+        """
+        Get the protein value for Pig and Poultry systems.
+
+        Returns:
+            float: The protein value in kg.
+        """
+        pig_and_poultry_protein = self.data_manager_class.get_static_livestock_protein_scaler(
+            year=self.baseline_year, system='Pig_Poultry',item='meat', abatement=self.abatement_type
+        )
+
+        return pig_and_poultry_protein["value"].item()
+    
+    def get_total_static_ag_protein(self):
+        """
+        Get the total protein value for all static agricultural systems.
+
+        Returns:
+            float: The total protein value in kg.
+        """
+        sheep_protein = self.get_sheep_protein()
+        pig_and_poultry_protein = self.get_pig_and_poultry_protein()
+
+        return sheep_protein + pig_and_poultry_protein