+++ tests/example.py
@@ -19,5 +19,9 @@ def main():
 
     print(optigob.get_total_land_area_by_sector_df())
 
+    print(optigob.get_total_protein_by_sector())
+
+    print(optigob.get_total_protein_by_sector_df())
+
 if __name__ == '__main__':
     main()