+++ tests/test_baseline_livestock.py
@@ -82,5 +82,15 @@ class TestBaselineLivestock(unittest.TestCase):
         result = self.baseline.get_total_area()
         print(f"total area: {result}")
 
+    def test_get_total_beef_protein(self):
+        result = self.baseline.get_total_beef_protein()
+        print(f"TOTAL beef protein: {result}")
+        self.assertIsNotNone(result)
+
+    def test_get_total_milk_protein(self):
+        result = self.baseline.get_total_milk_protein()
+        print(f"TOTAL milk protein: {result}")
+        self.assertIsNotNone(result)
+
 if __name__ == '__main__':
     unittest.main()